# lib/api-spec/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "Api",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "paths": {
...
    "/admin/overview": {
      "get": {
        "operationId": "getAdminOverview",
        "tags": [
          "ecomstack"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOverview"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Sign in required"
          },
          "403": {
            "description": "Access denied"
          }
        }
      }
    },
    "/admin/resources": {
      "get": {
        "operationId": "listAdminResources",
        "tags": [
          "ecomstack"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Resource"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Sign in required"
          },
          "403": {
            "description": "Access denied"
          }
        }
      },
      "post": {
        "operationId": "createResource",
        "tags": [
          "ecomstack"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResourceInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminResource"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Sign in required"
          },
          "403": {
            "description": "Access denied"
          }
        }
      }
    },
    "/admin/resources/{id}": {
      "get": {
        "operationId": "getAdminResource",
        "tags": [
          "ecomstack"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminResource"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Sign in required"
          },
          "403": {
            "description": "Access denied"
          }
        }
      },
      "put": {
        "operationId": "updateResource",
        "tags": [
          "ecomstack"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResourceInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminResource"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Sign in required"
          },
          "403": {
            "description": "Access denied"
          }
        }
      }
    },
    "/admin/users": {
      "get": {
        "operationId": "listUsers",
        "tags": [
          "ecomstack"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserSummary"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Sign in required"
          },
          "403": {
            "description": "Access denied"
          }
        }
      }
    },
    "/admin/access": {
      "post": {
        "operationId": "changeAccess",
        "tags": [
          "ecomstack"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GrantInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Sign in required"
          },
          "403": {
            "description": "Access denied"
          }
        }
      }
    },
    "/admin/support": {
      "get": {
        "operationId": "listSupportRequests",
        "tags": [
          "ecomstack"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SupportRequest"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Sign in required"
          },
          "403": {
            "description": "Access denied"
          }
        }
      }
    },
    "/admin/support/{id}": {
      "patch": {
        "operationId": "updateSupportRequest",
        "tags": [
          "ecomstack"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SupportUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Sign in required"
          },
          "403": {
            "description": "Access denied"
          }
        }
      }
    },
    "/admin/settings": {
      "get": {
        "operationId": "getAdminSettings",
        "tags": [
          "ecomstack"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminSettings"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Sign in required"
          },
          "403": {
            "description": "Access denied"
          }
        }
      },
      "put": {
        "operationId": "updateSettings",
        "tags": [
          "ecomstack"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminSettings"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Sign in required"
          },
          "403": {
            "description": "Access denied"
          }
        }
      }
    },
    "/admin/impact/status": {
      "get": {
        "operationId": "getSyncStatus",
        "tags": [
          "ecomstack"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncStatus"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Sign in required"
          },
          "403": {
            "description": "Access denied"
          }
        }
      }
    },
    "/admin/impact/diagnostic": {
      "post": {
        "operationId": "runImpactDiagnostic",
        "tags": [
          "ecomstack"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncStatus"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Sign in required"
          },
          "403": {
            "description": "Access denied"
          }
        }
      }
    },
    "/admin/impact/sync": {
      "post": {
        "operationId": "runImpactSync",
        "tags": [
          "ecomstack"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncStatus"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Sign in required"
          },
          "403": {
            "description": "Access denied"
          }
        }
      }
    },
    "/admin/taxonomies": {
      "get": {
        "operationId": "listTaxonomies",
        "tags": [
          "ecomstack"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Taxonomy"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Sign in required"
          },
          "403": {
            "description": "Access denied"
          }
        }
      },
      "post": {
        "operationId": "createTaxonomy",
        "tags": [
          "ecomstack"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaxonomyInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Taxonomy"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Sign in required"
          },
          "403": {
            "description": "Access denied"
          }
        }
      }
    },
    "/admin/taxonomies/{id}": {
      "put": {
        "operationId": "updateTaxonomy",
        "tags": [
          "ecomstack"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaxonomyInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Taxonomy"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Sign in required"
          },
          "403": {
            "description": "Access denied"
          }
        }
      },
      "delete": {
        "operationId": "deleteTaxonomy",
        "tags": [
          "ecomstack"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Sign in required"
          },
          "403": {
            "description": "Access denied"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Message": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ]
      },
...
      "Resource": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "Prompt",
              "Skill",
              "Cheat Sheet"
            ]
          },
          "category": {
            "type": "string"
          },
          "tool": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "preview": {
            "type": "string"
          },
          "useCase": {
            "type": "string"
          },
          "instructions": {
            "type": "string"
          },
          "tutorialUrl": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "isFree": {
            "type": "boolean"
          },
          "featured": {
            "type": "boolean"
          },
          "isDemo": {
            "type": "boolean"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "published",
              "archived"
            ]
          },
          "updatedAt": {
            "type": "string"
          },
          "coverUrl": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "slug",
          "title",
          "description",
          "type",
          "category",
          "tool",
          "tags",
          "preview",
          "useCase",
          "instructions",
          "tutorialUrl",
          "version",
          "isFree",
          "featured",
          "isDemo",
          "status",
          "updatedAt",
          "coverUrl"
        ]
      },
      "ResourceInput": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "Prompt",
              "Skill",
              "Cheat Sheet"
            ]
          },
          "category": {
            "type": "string"
          },
          "tool": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "preview": {
            "type": "string"
          },
          "useCase": {
            "type": "string"
          },
          "instructions": {
            "type": "string"
          },
          "tutorialUrl": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "isFree": {
            "type": "boolean"
          },
          "featured": {
            "type": "boolean"
          },
          "isDemo": {
            "type": "boolean"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "published",
              "archived"
            ]
          },
          "coverUrl": {
            "type": "string"
          },
          "content": {
            "type": "string"
          }
        },
        "required": [
          "slug",
          "title",
          "description",
          "type",
          "category",
          "tool",
          "preview",
          "content",
          "status",
          "isFree"
        ]
      },
...
      "SupportUpdate": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "open",
              "resolved"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
...
      "GrantInput": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "action": {
            "type": "string",
            "enum": [
              "grant",
              "revoke"
            ]
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "userId",
          "action",
          "reason"
        ]
      },
...
      "AdminOverview": {
        "type": "object",
        "properties": {
          "registeredUsers": {
            "type": "number"
          },
          "signupClicks": {
            "type": "number"
          },
          "onboardingCompletions": {
            "type": "number"
          },
          "verifiedPaidTrials": {
            "type": "number"
          },
          "accountsWithAccess": {
            "type": "number"
          },
          "resourceCopies": {
            "type": "number"
          },
          "resourceDownloads": {
            "type": "number"
          },
          "resources": {
            "type": "number"
          },
          "openSupportRequests": {
            "type": "number"
          },
          "audit": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Audit"
            }
          }
        },
        "required": [
          "registeredUsers",
          "signupClicks",
          "verifiedPaidTrials",
          "accountsWithAccess",
          "onboardingCompletions",
          "resourceCopies",
          "resourceDownloads",
          "resources",
          "openSupportRequests",
          "audit"
        ]
      },
      "AdminResource": {
        "type": "object",
        "properties": {
          "resource": {
            "$ref": "#/components/schemas/Resource"
          },
          "content": {
            "type": "string"
          },
          "assets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Asset"
            }
          }
        },
        "required": [
          "resource",
          "content",
          "assets"
        ]
      },
...
    }
  }
}